split day night switch functions to separated files, move day night switch to composable
